Modularise the element create() method for tests.
