+++ test/lib/create.js
@@ -1,0 +1,9 @@
+export default function (name, attrs = {}, chren = []) {
+  const elem = document.createElement(name);
+  const attrsIsChren = Array.isArray(attrs);
+  chren = attrsIsChren ? attrs : chren;
+  attrs = attrsIsChren ? {} : attrs;
+  Object.keys(attrs).forEach(key => elem.setAttribute(key, attrs[key]));
+  chren.forEach(ch => elem.appendChild(ch));
+  return elem;
+}
