+++ test/unit/internal/get-slot.js
@@ -1,14 +1,5 @@
+import create from '../../lib/create';
 import getSlot from '../../../src/internal/get-slot';
-
-function create (name, attrs = {}, chren = []) {
-  const elem = document.createElement(name);
-  const attrsIsChren = Array.isArray(attrs);
-  chren = attrsIsChren ? attrs : chren;
-  attrs = attrsIsChren ? {} : attrs;
-  Object.keys(attrs).forEach(key => elem.setAttribute(key, attrs[key]));
-  chren.forEach(ch => elem.appendChild(ch));
-  return elem;
-}
 
 describe('internal/get-slot', function () {
   it('returns a <slot>', function () {
